refactor(modularization): extract specific logic and API concerns into dedicated modules
