+++ src/shared/utils/normalizeContainerNumber.ts
@@ -1,0 +1,3 @@
+export function normalizeContainerNumber(value: string): string {
+  return value.trim().toUpperCase()
+}
